move contrastive CLI to its own module
